test(vetting): put reference tables with the assertions that read them
SPEC 6.3.1 said a table could sit in a _common.h and take the oracle of whoever
read it. Review asked for the opposite rule, and the opposite rule is right: a
shared table is read by callers it cannot see, which is how three separate
_<oracle> tests ended up judging themselves against Nyxus snapshots. 6.3.1 now
reads "a table lives in the file whose assertions read it; _common.h holds
fixtures, never values", plus the corollary that a helper serving several oracle
files means one table answering several claims, so the helper splits per oracle.

Three defects the split exposed, all the same shape as the GLCM matlab one:

- test_3d_morphology_mesh_volume_matlab asserted against
  morphology_3d_regression_ref_vals (478516, Nyxus) and never touched MATLAB's
  497824. The one assertion carrying that matlab claim could not have failed on a
  MATLAB disagreement. It now reads morphology_3d_matlab_ref_vals with the
  measured per-feature band, and passes at 3.58%.
- shape2d_ref_val searched matlab, skimage, imea and regression tables in one
  pass, so a _matlab assertion could resolve its "oracle" value out of a snapshot.
  Split into per-oracle helpers.
- test_2d_morphology_skimage.h asserted CIRCULARITY against the regression
  snapshot, duplicating test_2d_morphology_analytic.h, which vets it against
  sqrt(4*pi*A)/P as the registry says. Removed rather than relocated.

Moves: morphology_3d_regression_ref_vals -> test_3d_morphology_regression.h, with
the shared helper reduced to a fixture that produces a value and judges nothing.
The six 2D shape tables -> their matlab / skimage / imea / cellprofiler /
regression files. The 119-row coverage baseline -> its six per-family
test_3d_<family>_coverage.h files, published into a registry because include order
puts the per-family files after the shared TEST_P bodies. CellProfiler's six keys
get their own table: unlike the GLCM case there is no evidence either way about
where those numbers came from, so the split preserves the registry's claim and
makes the missing SPEC 6.4 provenance local instead of hidden behind a shared
table.

New test_2d_neighbor_common.h holds the neighbour fixture builder, so neither
oracle file has to include test_2d_neighbor_regression.h to borrow scaffolding.

Still to come: test_2d_remaining_common.h (imea tables renamed by benchmark and
moved with their 19 assertions, chords and polygonality to regression, zernike to
its own file), the radial table's home, and the checker rule that rejects a
reference table declared in a _common.h.
