+++ tests/test_2d_morphology_common.h
@@ -22,140 +22,6 @@
 #include <filesystem>
 #include <memory>
 #include "../src/nyx/grayscale_tiff.h"
-#include "test_ref_vals.h"
-
-// "Unvetted" means this table is not claiming an accepted independent
-// built-in/package oracle for these rows. Some entries, such as area,
-// centroids, and bounding-box limits, are easy to inspect on this small
-// fixture; that kind of hand recomputation is useful as a sanity check but is
-// not V&V by our tracker definition, because it restates the fixture geometry
-// and Nyxus coordinate/spacing conventions rather than using a third-party
-// oracle. Rows with accepted external built-in/API comparators live in the
-// oracle_3p table below.
-static ref_vals_map<double> morphology_2d_regression_ref_vals{
-	{"AREA_PIXELS_COUNT", 26.0},
-	{"AREA_UM2", 104.0},
-	{"CENTROID_X", 2.61538461538462},
-	{"CENTROID_Y", 2.84615384615385},
-	{"WEIGHTED_CENTROID_X", 2.84160305343511},
-	{"WEIGHTED_CENTROID_Y", 3.43893129770992},
-	{"MASS_DISPLACEMENT", 0.634476074243407},
-	{"COMPACTNESS", 0.027517678630878},
-	{"BBOX_XMIN", 0.0},
-	{"BBOX_YMIN", 0.0},
-	{"BBOX_WIDTH", 6.0},
-	{"BBOX_HEIGHT", 7.0},
-	{"DIAMETER_EQUAL_AREA", 5.75362739175159},
-	{"EXTENT", 0.619047619047619},
-	{"ASPECT_RATIO", 0.857142857142857},
-	{"MAJOR_AXIS_LENGTH", 6.96881616898619},
-	{"MINOR_AXIS_LENGTH", 5.48870991295738},
-	{"ELONGATION", 0.787610087547462},
-	{"ECCENTRICITY", 0.616173960820708},
-	{"ORIENTATION", 70.4173944984207},
-	{"ROUNDNESS", 0.681656295209303},
-	{"PERIMETER", 26.9349412836191},
-	{"EDGE_MEAN_INTENSITY", 41.8333333333333},
-	{"EDGE_STDDEV_INTENSITY", 16.7691944455582},
-	{"EDGE_MAX_INTENSITY", 68.0},
-	{"EDGE_MIN_INTENSITY", 12.0},
-	{"EDGE_INTEGRATED_INTENSITY", 753.0},
-	{"CIRCULARITY", 0.671081973229055},
-	{"EULER_NUMBER", 0.0},
-	{"DIAMETER_MIN_ENCLOSING_CIRCLE", 6.32475519180298},
-	{"ROI_RADIUS_MEAN", 1.07692307692308},
-	{"ROI_RADIUS_MAX", 4.0},
-	{"ROI_RADIUS_MEDIAN", 1.0},
-};
-
-// Fractal dimensions are validated on tests/data/fractal_blob512_seg.ome.tif, a large (512x512)
-// irregular ROI. Offline ImageJ/FracLac oracles:
-//   FRACT_DIM_BOXCOUNT  = box count of the filled ROI (single origin-aligned grid, the same method
-//                         as Nyxus' large-ROI path) -> 1.8706, asserted at 1%.
-//   FRACT_DIM_PERIMETER = box count of the ROI edge (cross-method vs Nyxus' Richardson divider,
-//                         estimating the same boundary dimension) -> 1.0493, asserted at 3%
-//                         (Nyxus divider = 1.0572, a 0.8% gap).
-static ref_vals_map<double> morphology_2d_fraclac_ref_vals{
-	{"FRACT_DIM_BOXCOUNT", 1.8706},   // vetted by ImageJ/FracLac: an independent implementation of the same box-count method
-	{"FRACT_DIM_PERIMETER", 1.0493},  // vetted by ImageJ/FracLac edge box-count: cross-method vs Nyxus' divider
-};
-
-// Extrema coordinates, vetted against MATLAB (registry: status=vetted, oracle=matlab).
-static ref_vals_map<double> morphology_2d_matlab_extrema_ref_vals{
-	// EXTREMA P1..P8 (X,Y) match MATLAB/Octave regionprops('Extrema') EXACTLY under the documented
-	// coordinate convention: MATLAB returns 1-based sub-pixel *corner* coords, Nyxus returns 0-based
-	// pixel *centers*. The corner is direction-specific -> the offset is per-point: left/top coords
-	// map as (matlab - 0.5), right/bottom coords as (matlab - 1.5). Verified on this 8x8 fixture by
-	// octave/matlab_oracle_tests/extrema_8x8.m: raw Extrema P1(2.5,0.5) P2(4.5,0.5) P3(6.5,2.5)
-	// P4(6.5,4.5) P5(5.5,7.5) P6(3.5,7.5) P7(0.5,4.5) P8(0.5,2.5) -> after the per-point offset ->
-	// P1(2,0) P2(3,0) P3(5,2) P4(5,3) P5(4,6) P6(3,6) P7(0,3) P8(0,2), i.e. these goldens exactly.
-	// (The earlier "~1px off" on the right/bottom coords was a harness bug: it used a uniform -0.5.)
-	{"EXTREMA_P1_X", 2.0},
-	{"EXTREMA_P1_Y", 0.0},
-	{"EXTREMA_P2_X", 3.0},
-	{"EXTREMA_P2_Y", 0.0},
-	{"EXTREMA_P3_X", 5.0},
-	{"EXTREMA_P3_Y", 2.0},
-	{"EXTREMA_P4_X", 5.0},
-	{"EXTREMA_P4_Y", 3.0},
-	{"EXTREMA_P5_X", 4.0},
-	{"EXTREMA_P5_Y", 6.0},
-	{"EXTREMA_P6_X", 3.0},
-	{"EXTREMA_P6_Y", 6.0},
-	{"EXTREMA_P7_X", 0.0},
-	{"EXTREMA_P7_Y", 3.0},
-	{"EXTREMA_P8_X", 0.0},
-	{"EXTREMA_P8_Y", 2.0},
-};
-
-// Convex-hull area, solidity and erosion count, vetted against scikit-image.
-static ref_vals_map<double> morphology_2d_skimage_hull_ref_vals{
-	// CONVEX_HULL_AREA / SOLIDITY are cross-checked against scikit-image on this exact ROI. Nyxus
-	// computes a Pick's-theorem pixel-count hull area (convex_hull_nontriv.cpp) = 27, solidity
-	// 26/27 = 0.9629630. Because Nyxus hulls through pixel CENTRES, this reproduces skimage's
-	// convex_hull_image(offset_coordinates=False) == 27 EXACTLY, so we vet against THAT convention
-	// (27 / 0.9629630) with a tight 1% tolerance (frac_tolerance=100). The hull area is a
-	// provably-exact integer lattice count, so 1% is float/platform slack -- not a convention fudge --
-	// and it still catches a >=1 px regression. (skimage's regionprops DEFAULT uses
-	// offset_coordinates=True, which first expands every pixel to its +/-0.5 corners and rasterises
-	// the hull to 28 / 0.9285714; that +1 px is a corner-expansion convention, not an error, and is
-	// why we pin the offset_coordinates=False value rather than the default.) SOLIDITY is thus a real
-	// skimage-vetted <= 1 check (unlike the old impossible 1.3), matched exactly rather than within a
-	// loose band.
-	{"CONVEX_HULL_AREA", 27.0},
-	{"SOLIDITY", 0.9629629629629629},
-	{"EROSIONS_2_VANISH", 1.0},
-};
-
-// Equal-perimeter diameter, geodetic length and thickness, vetted against imea.
-static ref_vals_map<double> morphology_2d_imea_geodetic_ref_vals{
-	{"DIAMETER_EQUAL_PERIMETER", 8.57365809435587},
-	// Real-valued rectangle-model roots after the geo_len_thickness.cpp perimeter-truncation fix
-	// (were the integer-truncated 10.0 / 3.0). Now vetted against imea's geodeticlength_and_thickness
-	// in test_2d_morphology_imea.h.
-	{"GEODETIC_LENGTH", 11.13182483477333},
-	{"THICKNESS", 2.3356458070362205},
-};
-
-// Circumscribing/inscribing circle diameters. status=regression: pinned Nyxus output with
-// no third-party oracle, so the name must not borrow the vetted claim of the tables above.
-static ref_vals_map<double> morphology_2d_regression_diameters_ref_vals{
-	{"DIAMETER_CIRCUMSCRIBING_CIRCLE", 12.3317073399088},
-	{"DIAMETER_INSCRIBING_CIRCLE", 0.828486893405308},
-};
-
-// These four were one table until SPEC 6.3.1 required a table to name a single oracle. One
-// lookup keeps the call site unchanged; each table now states where its own numbers came from.
-static bool shape2d_ref_val (const std::string& key, double& out)
-{
-    for (const auto* t : { &morphology_2d_matlab_extrema_ref_vals, &morphology_2d_skimage_hull_ref_vals, &morphology_2d_imea_geodetic_ref_vals, &morphology_2d_regression_diameters_ref_vals })
-    {
-        auto it = t->find(key);
-        if (it != t->end()) { out = it->second; return true; }
-    }
-    return false;
-}
-
 
 static Fsettings make_shape2d_settings()
 {
@@ -233,28 +99,7 @@ static void calculate_shape2d_feature_values(std::vector<std::vector<double>>& f
 	fvals = roidata.fvals;
 }
 
-static void assert_unvetted_no_direct_oracle_shape2d_feature(
-	const std::vector<std::vector<double>>& fvals,
-	Nyxus::Feature2D feature,
-	const std::string& feature_name,
-	double frac_tolerance = 1000.0)
-{
-	SCOPED_TRACE(std::string("UNVETTED_NO_DIRECT_ORACLE__") + feature_name);
-	ASSERT_TRUE(morphology_2d_regression_ref_vals.count(feature_name) > 0);
-	ASSERT_TRUE(agrees_gt(fvals[static_cast<int>(feature)][0], morphology_2d_regression_ref_vals[feature_name], frac_tolerance));
-}
 
-static void assert_verifiable_with_3p_builtin_oracle_shape2d_feature(
-	const std::vector<std::vector<double>>& fvals,
-	Nyxus::Feature2D feature,
-	const std::string& feature_name,
-	double frac_tolerance = 1000.0)
-{
-	SCOPED_TRACE(std::string("VERIFIABLE_WITH_3P_BUILTIN_ORACLE__") + feature_name);
-	double ref_val{};
-	ASSERT_TRUE(shape2d_ref_val(feature_name, ref_val));
-	ASSERT_TRUE(agrees_gt(fvals[static_cast<int>(feature)][0], ref_val, frac_tolerance));
-}
 
 // Loads the large ROI mask tests/data/fractal_blob512_seg.ome.tif (path resolved relative to this
 // source file) into a single-ROI LR and computes the fractal features.
@@ -301,3 +146,11 @@ static void calculate_fractal_blob512_feature_values(std::vector<std::vector<dou
 
 	fvals = roidata.fvals;
 }
+
+// Fixtures only, no reference data and no assertions (SPEC 6.3.1). The six tables that used to live
+// here now sit with the assertions that read them, and the two shared assert helpers are split per
+// oracle. They had to be: shape2d_ref_val() searched the matlab, skimage, imea and regression tables
+// in one pass, so a _matlab assertion could resolve its "oracle" value out of a snapshot, and
+// assert_unvetted_no_direct_oracle_shape2d_feature() judged _skimage and _cellprofiler functions
+// against morphology_2d_regression_ref_vals -- the SPEC 6.2.1 defect, reached through file layout
+// rather than through naming.
