+++ tests/test_2d_neighbor_common.h
@@ -1,0 +1,80 @@
+#pragma once
+
+// Shared fixture for the 2D neighbour-graph tests: the settings recipe and the scene builder that
+// turns neighborhood2d_scene_labels into a populated roiData map with basic-morphology, contour and
+// neighbour features computed.
+//
+// Fixtures only, no reference data (SPEC 6.3.1). This header exists because both oracle files
+// (test_2d_neighbor_cellprofiler.h, test_2d_neighbor_analytic.h) previously reached the builder by
+// including test_2d_neighbor_regression.h -- an oracle file including a regression file to borrow
+// scaffolding, which put a snapshot table in scope of every assertion that did so.
+
+#include <gtest/gtest.h>
+
+#include <string>
+#include <unordered_map>
+#include <unordered_set>
+
+#include "../src/nyx/roi_cache.h"
+#include "../src/nyx/features/basic_morphology.h"
+#include "../src/nyx/features/contour.h"
+#include "../src/nyx/features/neighbors.h"
+#include "test_data.h"
+#include "test_main_nyxus.h"
+
+static Fsettings make_neighbors2d_settings()
+{
+	Fsettings s;
+	s.resize(static_cast<int>(NyxSetting::__COUNT__));
+	s[static_cast<int>(NyxSetting::SOFTNAN)].rval = 0.0;
+	s[static_cast<int>(NyxSetting::TINY)].rval = 0.0;
+	s[static_cast<int>(NyxSetting::SINGLEROI)].bval = false;
+	s[static_cast<int>(NyxSetting::GREYDEPTH)].ival = 128;
+	s[static_cast<int>(NyxSetting::PIXELSIZEUM)].rval = 1.0;
+	s[static_cast<int>(NyxSetting::XYRES)].rval = 1.0;
+	s[static_cast<int>(NyxSetting::PIXELDISTANCE)].ival = 1;
+	s[static_cast<int>(NyxSetting::USEGPU)].bval = false;
+	s[static_cast<int>(NyxSetting::VERBOSLVL)].ival = 0;
+	s[static_cast<int>(NyxSetting::IBSI)].bval = false;
+	return s;
+}
+
+static void calculate_neighbor_feature_values(std::unordered_map<int, LR>& roiData)
+{
+	Fsettings s = make_neighbors2d_settings();
+	std::unordered_set<int> uniqueLabels;
+
+	for (const auto& px : neighborhood2d_scene_labels)
+	{
+		int label = static_cast<int>(px.intensity);
+		uniqueLabels.insert(label);
+
+		auto [it, inserted] = roiData.try_emplace(label, label);
+		LR& roi = it->second;
+
+		if (inserted)
+			init_label_record_3(roi, static_cast<int>(px.x), static_cast<int>(px.y), 1);
+		else
+			update_label_record_3(roi, static_cast<int>(px.x), static_cast<int>(px.y), 1);
+
+		roi.raw_pixels.push_back(Pixel2(static_cast<size_t>(px.x), static_cast<size_t>(px.y), static_cast<PixIntens>(1)));
+	}
+
+	BasicMorphologyFeatures basic;
+	ContourFeature contour;
+	for (auto& item : roiData)
+	{
+		LR& roi = item.second;
+		roi.make_nonanisotropic_aabb();
+		roi.aux_image_matrix = ImageMatrix(roi.raw_pixels);
+		roi.initialize_fvals();
+
+		basic.calculate(roi, s);
+		basic.save_value(roi.fvals);
+
+		contour.calculate(roi, s);
+		contour.save_value(roi.fvals);
+	}
+
+	NeighborsFeature::manual_reduce(roiData, s, uniqueLabels);
+}
