+++ tests/test_3d_glcm_coverage.h
@@ -19,3 +19,49 @@ INSTANTIATE_TEST_SUITE_P(
 	Test3DFeature_UNVETTED_LOCAL_REGRESSION,
 	testing::ValuesIn(feature_3d_cases_for_family("glcm", false)),
 	sanitize_3d_feature_test_name);
+
+// Regression baselines for this family's slice of the sweep: pinned Nyxus output for the public 3D
+// glcm features that no third-party oracle backs yet. They establish no vetting (SPEC 1), and they
+// live here rather than in the shared harness so the table sits with the assertions that read it.
+static ref_vals_map<std::vector<double>> glcm_3d_regression_coverage_ref_vals
+{
+	{ "3GLCM_ACOR_AVE", { 896.29490954682888 } },
+	{ "3GLCM_ASM_AVE", { 0.21714923037245615 } },
+	{ "3GLCM_CLUPROM_AVE", { 5319702.1464416385 } },
+	{ "3GLCM_CLUSHADE_AVE", { 25300.544725496744 } },
+	{ "3GLCM_CLUTEND_AVE", { 1830.7699876860852 } },
+	{ "3GLCM_CONTRAST_AVE", { 190.39105377069197 } },
+	{ "3GLCM_CORRELATION_AVE", { 0.81160543314537192 } },
+	{ "3GLCM_DIFAVE_AVE", { 4.4826978263468851 } },
+	{ "3GLCM_DIFENTRO_AVE", { 2.5804645841819687 } },
+	{ "3GLCM_DIFVAR_AVE", { 168.82038645655439 } },
+	{ "3GLCM_DIS", { 5.5662607942292563, 5.2865125777406536, 5.5662607942292563, 3.5205463112065347, 3.2058935920047036, 3.5205463112065347, 5.5662607942292563, 5.2865125777406536, 5.5662607942292563, 4.6920801473460996, 4.4260407115349301, 4.6920801473460996, 1.379816189466244 } },
+	{ "3GLCM_DIS_AVE", { 4.4826978263468824 } },
+	{ "3GLCM_ENERGY", { 0.20619709037207257, 0.21210950671886142, 0.20619709037207257, 0.22347189327232553, 0.229768029906674, 0.22347189327232553, 0.20619709037207257, 0.21210950671886142, 0.20619709037207257, 0.21657166507454639, 0.22249608735556295, 0.21657166507454639, 0.2415813859599367 } },
+	{ "3GLCM_ENERGY_AVE", { 0.21714923037245615 } },
+	{ "3GLCM_ENTROPY", { 6.1375193094015223, 6.0251268596431471, 6.1375193094015223, 5.6356261936885357, 5.5229037686481623, 5.6356261936885357, 6.1375193094015223, 6.0251268596431471, 6.1375193094015223, 5.9202631504971848, 5.8126798565436344, 5.9202631504971848, 4.8177837873733909 } },  // FIX: were buggy unnormalized ~-6.8e6; post /sum_p fix (3d_glcm.cpp)
+	{ "3GLCM_ENTROPY_AVE", { 5.8358059275253087 } },  // FIX: was buggy unnormalized -7.09e6; post /sum_p fix
+	{ "3GLCM_HOM1", { 0.63159286085358357, 0.67511464571418467, 0.63159286085358357, 0.675209797165082, 0.7178512712427948, 0.675209797165082, 0.63159286085358357, 0.67511464571418467, 0.63159286085358357, 0.65283716836970751, 0.69041197263012255, 0.65283716836970751, 0.77822827662266392 } },
+	{ "3GLCM_HOM1_AVE", { 0.67070662972368189 } },
+	{ "3GLCM_HOM2", { 0.59196175760722125, 0.64232872280698583, 0.59196175760722125, 0.63506326531940249, 0.68470533035896952, 0.63506326531940249, 0.59196175760722125, 0.64232872280698583, 0.59196175760722125, 0.61310667832992871, 0.65700403974706201, 0.61310667832992871, 0.75964624124070679 } },  // FIX: were buggy unnormalized ~3.1e5; post /sum_p fix (homogeneity now in [0,1])
+	{ "3GLCM_IDMN_AVE", { 0.97393106857854816 } },
+	{ "3GLCM_IDM_AVE", { 0.63463076728371071 } },
+	{ "3GLCM_IDN_AVE", { 0.95491658377679667 } },
+	{ "3GLCM_ID_AVE", { 0.67070662972368211 } },
+	{ "3GLCM_INFOMEAS1_AVE", { -0.44521529228752316 } },
+	{ "3GLCM_INFOMEAS2_AVE", { 0.97909293489987548 } },
+	{ "3GLCM_IV_AVE", { 0.15630830580149896 } },
+	{ "3GLCM_JAVE_AVE", { 22.049501181077645 } },
+	{ "3GLCM_JE_AVE", { 5.83580592757035 } },
+	{ "3GLCM_JMAX_AVE", { 0.46494916987187934 } },
+	{ "3GLCM_JVAR_AVE", { 505.29026036419413 } },
+	{ "3GLCM_SUMAVERAGE_AVE", { 44.099002362155289 } },
+	{ "3GLCM_SUMENTROPY_AVE", { 4.3772246057690438 } },
+	{ "3GLCM_SUMVARIANCE", { 1783.3177119109412, 1796.9249632904919, 1783.3177119109412, 1872.6626392475214, 1887.8346987420691, 1872.6626392475214, 1783.3177119109412, 1796.9249632904919, 1783.3177119109412, 1820.9267907064923, 1833.7216725021183, 1820.9267907064923, 1964.1538345421473 } },
+	{ "3GLCM_SUMVARIANCE_AVE", { 1830.7699876860852 } },
+	{ "3GLCM_VARIANCE", { 504.75832684863678, 506.44120833399279, 504.75832684863678, 504.89933309869377, 506.49086530867982, 504.89933309869377, 504.75832684863678, 506.44120833399279, 504.75832684863678, 504.73432274406343, 506.30964057188811, 504.73432274406343, 504.78984310590351 } },
+	{ "3GLCM_VARIANCE_AVE", { 505.29026036419378 } },
+};
+
+static const bool glcm_3d_coverage_baseline_registered =
+	register_coverage_baseline("glcm", &glcm_3d_regression_coverage_ref_vals);
