+++ tests/test_all.cc
@@ -24,6 +24,7 @@
 #include "test_2d_moments_regression.h"
 #include "test_2d_remaining_common.h"
 #include "test_2d_zernike_regression.h"
+#include "test_2d_neighbor_common.h"
 #include "test_2d_neighbor_regression.h"
 #include "test_2d_neighbor_cellprofiler.h"
 #include "test_2d_neighbor_analytic.h"
